--- deps/build_patched_wasme.sh
@@ -1,5 +1,0 @@
-#!/bin/bash
-cd wasm/tools/wasme/cli
-make _output/wasme
-cp _output/wasme ../../../../bin/wasme_patched
-cd -
