--- latency_monitoring/Makefile
@@ -1,2 +1,0 @@
-hellomake: parallel_curl.c
-	gcc parallel_curl.c -lpthread -o pc
