--- latency_monitoring/build_patched_wasme.sh
@@ -1,8 +1,0 @@
-#!/bin/bash
-cd wasm/tools/wasme/cli
-make _output/wasme
-if [ $? -eq 0 ]; then
-    mv _output/wasme ../../../../bin/wasme_patched
-fi
-
-cd -
