--- latency_monitoring/fault-injection.yaml
@@ -1,17 +1,0 @@
-apiVersion: networking.istio.io/v1alpha3
-kind: VirtualService
-metadata:
-  name: ratings
-spec:
-  hosts:
-  - ratings
-  http:
-  - route:
-    - destination:
-        host: ratings
-        subset: v1
-    fault:
-      delay:
-        percentage:
-          value: 100.0
-        fixedDelay: 50ms
