--- latency_monitoring/latency_monitor/.gitignore
@@ -1,1 +1,0 @@
-bazel-*
