--- latency_monitoring/latency_monitor/bazel/external/emscripten-toolchain.BUILD
@@ -1,6 +1,0 @@
-package(default_visibility = ['//visibility:public'])
-
-filegroup(
-  name = "all",
-  srcs = glob(["**/*"]),
-)
