--- latency_monitoring/latency_monitor/build_script.sh
@@ -1,7 +1,0 @@
-#!/bin/bash
-bazel build //:filter.wasm
-~/.wasme/bin/wasme build precompiled ./bazel-bin/filter.wasm --tag webassemblyhub.io/fruffy/test-filter:1  --config runtime-config.json
-# wasme build cpp -t test-filter . --tag webassemblyhub.io/fruffy/test-filter:1 --config runtime-config.json
-# wasme push webassemblyhub.io/fruffy/test-filter:1
-
-
