--- latency_monitoring/latency_monitor/filter.proto
@@ -1,5 +1,0 @@
-syntax = "proto3";
-
-message Config {
-    string value = 1;
-}
