--- latency_monitoring/latency_monitor/toolchain/cc_toolchain_config.bzl
@@ -1,100 +1,0 @@
-load("@bazel_tools//tools/cpp:cc_toolchain_config_lib.bzl",
-     "feature",
-     "flag_group",
-     "flag_set",
-     "tool_path")
-load("@bazel_tools//tools/build_defs/cc:action_names.bzl", "ACTION_NAMES")
-
-
-def _impl(ctx):
-    tool_paths = [
-        tool_path(
-            name = "gcc",
-            path = "emcc.sh",
-        ),
-        tool_path(
-            name = "ld",
-            path = "emcc.sh",
-        ),
-        tool_path(
-            name = "ar",
-            path = "emar.sh",
-        ),
-        tool_path(
-            name = "cpp",
-            path = "/bin/false",
-        ),
-        tool_path(
-            name = "gcov",
-            path = "/bin/false",
-        ),
-        tool_path(
-            name = "nm",
-            path = "/bin/false",
-        ),
-        tool_path(
-            name = "objdump",
-            path = "/bin/false",
-        ),
-        tool_path(
-            name = "strip",
-            path = "/bin/false",
-        ),
-    ]
-
-    toolchain_include_directories_feature = feature(
-        name = "toolchain_include_directories",
-        enabled = True,
-        flag_sets = [
-            flag_set(
-                actions = [
-                    ACTION_NAMES.assemble,
-                    ACTION_NAMES.preprocess_assemble,
-                    ACTION_NAMES.linkstamp_compile,
-                    ACTION_NAMES.c_compile,
-                    ACTION_NAMES.cpp_compile,
-                    ACTION_NAMES.cpp_header_parsing,
-                    ACTION_NAMES.cpp_module_compile,
-                    ACTION_NAMES.cpp_module_codegen,
-                    ACTION_NAMES.lto_backend,
-                    ACTION_NAMES.clif_match,
-                ],
-                flag_groups = [
-                    flag_group(
-                        flags = [
-                            "-isystem",
-                            "external/emscripten_toolchain/upstream/emscripten/system/include/libcxx",
-                            "-isystem",
-                            "external/emscripten_toolchain/upstream/emscripten/system/include/libc",
-                        ],
-                    ),
-                ],
-            ),
-        ],
-    )
-
-    return cc_common.create_cc_toolchain_config_info(
-        ctx = ctx,
-        toolchain_identifier = "wasm-toolchain",
-        host_system_name = "i686-unknown-linux-gnu",
-        target_system_name = "wasm32-unknown-emscripten",
-        target_cpu = "wasm",
-        target_libc = "unknown",
-        compiler = "emscripten",
-        abi_version = "unknown",
-        abi_libc_version = "unknown",
-        tool_paths = tool_paths,
-        # we don't need to use features, as emcc already adds the directories.
-        # we just need to include them here so that bazel doesn't complain on 
-        # "this rule is missing dependency declarations for the following files included".
-        cxx_builtin_include_directories = ["external/emscripten_toolchain/upstream/emscripten/system/include/libcxx",
-        "external/emscripten_toolchain/upstream/emscripten/system/include/libc",
-        ]
-        # features = [toolchain_include_directories_feature],
-    )
-
-cc_toolchain_config = rule(
-    implementation = _impl,
-    attrs = {},
-    provides = [CcToolchainConfigInfo],
-)
