--- latency_monitoring/latency_monitor/toolchain/emar.sh
@@ -1,6 +1,0 @@
-#!/bin/bash
-set -euo pipefail
-
-. $(dirname $0)/common.sh
-
-emar "$@"
