+++ latency_monitoring/message_counter/bazel-message_counter
@@ -1,0 +1,1 @@
+/home/fruffy/.cache/bazel/_bazel_fruffy/10cf5a69a52464f9bf69dfce5ab49ad8/execroot/filter_example
