+++ latency_monitoring/message_counter/filter.cc
@@ -1,193 +1,267 @@
-// NOLINT(namespace-envoy)
 #include <string>
 #include <unordered_map>
 
+#include "filter.pb.h"
 #include "google/protobuf/util/json_util.h"
 #include "proxy_wasm_intrinsics.h"
-#include "filter.pb.h"
+
+#define REQUEST_THRESHOLD 10
+#define STORAGE_NAME "storage-upstream"
 
 enum class TrafficDirection : int64_t {
-  Unspecified = 0,
-  Inbound = 1,
-  Outbound = 2,
+    Unspecified = 0,
+    Inbound = 1,
+    Outbound = 2,
 };
 
 // Retrieves the traffic direction from the configuration context.
 TrafficDirection getTrafficDirection() {
-  int64_t direction;
-  if (getValue({"listener_direction"}, &direction)) {
-    return static_cast<TrafficDirection>(direction);
-  }
-  return TrafficDirection::Unspecified;
+    int64_t direction;
+    if (getValue({"listener_direction"}, &direction)) {
+        return static_cast<TrafficDirection>(direction);
+    }
+    return TrafficDirection::Unspecified;
 }
 
 std::string trafficDirectionToString(TrafficDirection dir) {
-  if (dir == TrafficDirection::Unspecified) {
-    return "unspecified";
-  } else if (dir == TrafficDirection::Inbound) {
-    return "inbound";
-  } else {
-    return "outbound";
-  }
+    if (dir == TrafficDirection::Unspecified) {
+        return "unspecified";
+    } else if (dir == TrafficDirection::Inbound) {
+        return "inbound";
+    } else {
+        return "outbound";
+    }
 }
 
 class BidiRootContext : public RootContext {
-public:
-  explicit BidiRootContext(uint32_t id, std::string_view root_id) : RootContext(id, root_id) {
-    std::string workload_name;
-    if (getValue({"node", "metadata", "WORKLOAD_NAME"}, &workload_name)) {
-      workload_name_ = workload_name;
-      std::string warning = "initialized workload_name: ";
-      warning = warning.append(workload_name_);
-      LOG_WARN(warning);
-    } else {
-      LOG_WARN("Failed to set workload name");
+ public:
+    explicit BidiRootContext(uint32_t id, std::string_view root_id)
+        : RootContext(id, root_id) {
+        std::string workload_name;
+        if (getValue({"node", "metadata", "WORKLOAD_NAME"}, &workload_name)) {
+            workload_name_ = workload_name;
+            std::string warning = "initialized workload_name: ";
+            warning = warning.append(workload_name_);
+            LOG_WARN(warning);
+        } else {
+            LOG_WARN("Failed to set workload name");
+        }
     }
-  }
-  bool onConfigure(size_t /* configuration_size */) override;
+    bool onConfigure(size_t /* configuration_size */) override;
 
-  bool onStart(size_t) override;
-  std::string_view getWorkloadName() { return workload_name_; }
-  void incrementCount() { count_++; }
-  void decrementCount() { count_--; }
-  int getCount() { return count_; }
+    bool onStart(size_t) override;
+    std::string_view getWorkloadName() { return workload_name_; }
+    void incrementCount() { count_++; }
+    void decrementCount() { count_--; }
+    int getCount() { return count_; }
+    std::string header_value_;
 
-  std::string header_value_;
-
-private:
-  std::string_view workload_name_;
-  int count_ = 0;
+ private:
+    std::string_view workload_name_;
+    int count_ = 0;
 };
 
 class BidiContext : public Context {
-public:
-  explicit BidiContext(uint32_t id, RootContext* root) : Context(id, root), root_(static_cast<BidiRootContext*>(static_cast<void*>(root))) {
-    direction_ = getTrafficDirection();
-    std::string warning = "Got traffic direction, is ";
-    warning = warning.append(trafficDirectionToString(direction_));
-    LOG_WARN(warning);
-  }
+ public:
+    explicit BidiContext(uint32_t id, RootContext *root)
+        : Context(id, root),
+          root_(static_cast<BidiRootContext *>(static_cast<void *>(root))) {
+        direction_ = getTrafficDirection();
+        LOG_WARN("Got traffic direction, is " +
+                 trafficDirectionToString(direction_));
+        gauge_ = Gauge<int>::New("queue_gauge", "queue_size");
+    }
 
-  void onCreate() override;
-  FilterHeadersStatus onRequestHeaders(uint32_t headers, bool end_of_stream) override;
-  FilterHeadersStatus onRequestHeadersInbound();
-  FilterHeadersStatus onRequestHeadersOutbound();
-  FilterDataStatus onRequestBody(size_t body_buffer_length, bool end_of_stream) override;
-  FilterHeadersStatus onResponseHeaders(uint32_t headers, bool end_of_stream) override;
-  FilterHeadersStatus onResponseHeadersInbound();
-  FilterHeadersStatus onResponseHeadersOutbound();
-  void onDone() override;
-  void onLog() override;
-  void onDelete() override;
-private:
+    void onCreate() override;
+    FilterHeadersStatus onRequestHeaders(uint32_t headers,
+                                         bool end_of_stream) override;
+    FilterHeadersStatus onRequestHeadersInbound();
+    FilterHeadersStatus onRequestHeadersOutbound();
+    FilterDataStatus onRequestBody(size_t body_buffer_length,
+                                   bool end_of_stream) override;
+    FilterDataStatus onResponseBody(size_t body_buffer_length,
+                                    bool end_of_stream) override;
+    FilterHeadersStatus onResponseHeaders(uint32_t headers,
+                                          bool end_of_stream) override;
+    FilterHeadersStatus onResponseHeadersInbound();
+    FilterHeadersStatus onResponseHeadersOutbound();
 
-  BidiRootContext* root_;
-  std::string b3_trace_id_;
-  std::string b3_span_id_;
-  std::string b3_parent_span_id_;
-  TrafficDirection direction_;
+    void onDone() override;
+    void onLog() override;
+    void onDelete() override;
+
+ private:
+    BidiRootContext *root_;
+    std::string b3_trace_id_;
+    std::string b3_span_id_;
+    std::string b3_parent_span_id_;
+    TrafficDirection direction_;
+    Gauge<int> *gauge_;
+    WasmResult store_warning();
+    void print_headers(WasmHeaderMapType type);
 };
 
-static RegisterContextFactory register_BidiContext(CONTEXT_FACTORY(BidiContext),
-                                                      ROOT_FACTORY(BidiRootContext),
-                                                      "add_header_root_id");
+static RegisterContextFactory
+    register_BidiContext(CONTEXT_FACTORY(BidiContext),
+                         ROOT_FACTORY(BidiRootContext), "bidi_root_id");
 
-bool BidiRootContext::onConfigure(size_t config_buffer_length) {
-  auto conf = getBufferBytes(WasmBufferType::PluginConfiguration, 0, config_buffer_length);
-  LOG_DEBUG("onConfigure " + conf->toString());
-  header_value_ = conf->toString();
-  return true; 
+WasmResult BidiContext::store_warning() {
+    std::string key = toString(getCurrentTimeNanoseconds());
+    std::string value = std::string(root_->getWorkloadName());
+
+    auto context_id = id();
+    auto callback = [context_id](uint32_t, size_t body_size, uint32_t) {
+        getContext(context_id)->setEffectiveContext();
+        auto body =
+            getBufferBytes(WasmBufferType::HttpCallResponseBody, 0, body_size);
+        LOG_WARN(std::string(body->view()));
+    };
+    auto result = root()->httpCall(STORAGE_NAME,
+                                   {{":method", "GET"},
+                                    {":path", "/store"},
+                                    {":authority", STORAGE_NAME},
+                                    {"key", key},
+                                    {"value", value}},
+                                   "", {}, 1000, callback);
+    if (result != WasmResult::Ok) {
+        LOG_ERROR("Failed to make a call to " + STORAGE_NAME + ": " +
+                  toString(result));
+    }
+    return result;
 }
 
-bool BidiRootContext::onStart(size_t) { LOG_DEBUG("onStart"); return true;}
+void BidiContext::print_headers(WasmHeaderMapType type) {
+    if (type == WasmHeaderMapType::RequestHeaders) {
+        auto result = getRequestHeaderPairs();
+        auto pairs = result->pairs();
+        LOG_WARN("Request headers: " + toString(pairs.size()));
+        for (auto &p : pairs) {
+            LOG_WARN(std::string(p.first) + " -> " + std::string(p.second));
+        }
+    } else if (type == WasmHeaderMapType::ResponseHeaders) {
+        auto result = getResponseHeaderPairs();
+        auto pairs = result->pairs();
+        LOG_WARN("Reponse headers: " + toString(pairs.size()));
+        for (auto &p : pairs) {
+            LOG_WARN(std::string(p.first) + " -> " + std::string(p.second));
+        }
+    }
+}
 
-void BidiContext::onCreate() { LOG_DEBUG(std::string("onCreate " + std::to_string(id()))); }
+////////////////// REQUEST HEADERS //////////////////
 
 FilterHeadersStatus BidiContext::onRequestHeaders(uint32_t, bool) {
-  LOG_DEBUG(std::string("onRequestHeaders ") + std::to_string(id()));
+    LOG_WARN("############################################");
+    // Print all request headers
+    print_headers(WasmHeaderMapType::RequestHeaders);
 
-  // Print all request headers
-  auto result = getRequestHeaderPairs();
-  auto pairs = result->pairs();
-  LOG_WARN(std::string("request headers: ") + std::to_string(pairs.size()));
-  for (auto &p : pairs) {
-      LOG_WARN(std::string(p.first) + std::string(" -> ") +
-               std::string(p.second));
-  }
-  int THRESHOLD = 0;
-  root_->incrementCount();
-  std::string warning = "incrementing count";
-  warning = warning.append(std::to_string(root_->getCount()));
-  LOG_WARN(warning);
-  if (root_->getCount() > THRESHOLD) {
-    LOG_DEBUG("above threshold");
-  }
-
-  replaceRequestHeader("x-envoy-force-trace", "true");
-  if (direction_ == TrafficDirection::Inbound) {
-    return onRequestHeadersInbound();
-  } else if (direction_ == TrafficDirection::Outbound) {
-    return onRequestHeadersOutbound();
-  }
-  LOG_WARN("didn't get direction in request header");
+    replaceRequestHeader("x-envoy-force-trace", "true");
+    if (direction_ == TrafficDirection::Inbound) {
+        return onRequestHeadersInbound();
+    } else if (direction_ == TrafficDirection::Outbound) {
+        return onRequestHeadersOutbound();
+    }
+    LOG_ERROR("Missing request header direction.");
+    return FilterHeadersStatus::Continue;
 }
 
 FilterHeadersStatus BidiContext::onRequestHeadersInbound() {
-  LOG_WARN("in on request headers inbound");
-  addResponseHeader("requestheaderINbound", "hi");
-  return FilterHeadersStatus::Continue;
+    LOG_WARN("Incrementing count.");
+    root_->incrementCount();
+    auto curr_count = root_->getCount();
+    gauge_->record(curr_count, 0);
+    LOG_WARN("Current count: " + toString(curr_count));
+    if (curr_count > REQUEST_THRESHOLD) {
+        LOG_ERROR("Request queue " + toString(curr_count) +
+                  "is above threshold " + toString(REQUEST_THRESHOLD) + ".");
+        store_warning();
+    }
+    return FilterHeadersStatus::Continue;
 }
 
 FilterHeadersStatus BidiContext::onRequestHeadersOutbound() {
-  LOG_WARN("in on request headers outbound");
-  addResponseHeader("requestheaderOUTbound", "hi");
-  return FilterHeadersStatus::Continue;
+    return FilterHeadersStatus::Continue;
 }
 
+////////////////// RESPONSE HEADERS //////////////////
+
+FilterHeadersStatus BidiContext::onResponseHeaders(uint32_t, bool) {
+    LOG_WARN("############################################");
+    print_headers(WasmHeaderMapType::ResponseHeaders);
+
+    replaceResponseHeader("location", "envoy-wasm");
+    replaceResponseHeader("x-envoy-force-trace", "true");
+    if (direction_ == TrafficDirection::Inbound) {
+        return onResponseHeadersInbound();
+    } else if (direction_ == TrafficDirection::Outbound) {
+        return onResponseHeadersOutbound();
+    }
+    LOG_WARN("in response headers but no direction given");
+    return FilterHeadersStatus::Continue;
+}
 
 FilterHeadersStatus BidiContext::onResponseHeadersInbound() {
-  LOG_WARN("in on response headers inbound");
-  addResponseHeader("responseheaderINbound", "hi");
-  return FilterHeadersStatus::Continue;
+    LOG_WARN("Decrementing count.");
+    root_->decrementCount();
+    auto curr_count = root_->getCount();
+    gauge_->record(curr_count, 0);
+    LOG_WARN("Current count: " + toString(curr_count));
+    return FilterHeadersStatus::Continue;
 }
 
 FilterHeadersStatus BidiContext::onResponseHeadersOutbound() {
-  LOG_WARN("in on response headers outbound");
-  addResponseHeader("responseheaderOUTbound", "hi");
-  return FilterHeadersStatus::Continue;
+    return FilterHeadersStatus::Continue;
 }
 
+////////////////// REQUEST BODY //////////////////
 
-FilterHeadersStatus BidiContext::onResponseHeaders(uint32_t, bool) {
-  LOG_DEBUG(std::string("onResponseHeaders ") + std::to_string(id()));
-  // Print all headers
-    auto result = getResponseHeaderPairs();
-    auto pairs = result->pairs();
-    LOG_WARN(std::string("response headers: ") + std::to_string(pairs.size()));
-    for (auto &p : pairs) {
-        LOG_WARN(std::string(p.first) + std::string(" -> ") +
-                 std::string(p.second));
-    }
-  std::string warning = "decrementing count";
-  warning = warning.append(std::to_string(root_->getCount()));
-  LOG_WARN(warning);
-  root_->decrementCount();
-  replaceResponseHeader("location", "envoy-wasm");
-  replaceResponseHeader("x-envoy-force-trace", "true");
-  if (direction_ == TrafficDirection::Inbound) {
-    return onResponseHeadersInbound();
-  } else if (direction_ == TrafficDirection::Outbound) {
-    return onResponseHeadersOutbound();
-  }
-  LOG_WARN("in response headers but no direction given");
+FilterDataStatus BidiContext::onRequestBody(size_t body_buffer_length,
+                                            bool end_of_stream) {
+    return FilterDataStatus::Continue;
 }
 
-FilterDataStatus BidiContext::onRequestBody(size_t body_buffer_length, bool end_of_stream) {
-  return FilterDataStatus::Continue;
+////////////////// RESPONSE BODY //////////////////
+
+FilterDataStatus BidiContext::onResponseBody(size_t body_buffer_length,
+                                             bool end_of_stream) {
+    return FilterDataStatus::Continue;
 }
 
-void BidiContext::onDone() { LOG_DEBUG(std::string("onDone " + std::to_string(id()))); }
+////////////////// CONFIGURE //////////////////
 
-void BidiContext::onLog() { LOG_DEBUG(std::string("onLog " + std::to_string(id()))); }
+bool BidiRootContext::onConfigure(size_t config_buffer_length) {
+    auto conf = getBufferBytes(WasmBufferType::PluginConfiguration, 0,
+                               config_buffer_length);
+    LOG_DEBUG("onConfigure " + conf->toString());
+    header_value_ = conf->toString();
+    return true;
+}
 
-void BidiContext::onDelete() { LOG_DEBUG(std::string("onDelete " + std::to_string(id()))); }
+////////////////// START //////////////////
+
+bool BidiRootContext::onStart(size_t) {
+    LOG_DEBUG("onStart");
+    return true;
+}
+
+////////////////// CREATE //////////////////
+
+void BidiContext::onCreate() {
+    LOG_DEBUG(std::string("onCreate " + toString(id())));
+}
+
+////////////////// DONE //////////////////
+
+void BidiContext::onDone() {
+    LOG_DEBUG(std::string("onDone " + toString(id())));
+}
+
+////////////////// LOG //////////////////
+
+void BidiContext::onLog() { LOG_DEBUG(std::string("onLog " + toString(id()))); }
+
+////////////////// DELETE //////////////////
+
+void BidiContext::onDelete() {
+    LOG_DEBUG(std::string("onDelete " + toString(id())));
+}
