+++ latency_monitoring/parallel_curl/Makefile
@@ -1,0 +1,2 @@
+hellomake: parallel_curl.c
+	gcc parallel_curl.c -lpthread -o pc
